+++ packages/unity-react-core/vite.config.js
@@ -31,7 +31,7 @@ export default defineConfig({
   },
   define: {
     process: { env: { NODE_ENV: process.env.NODE_ENV } },
-    global: {},
+    global: "globalThis",
   },
   plugins: [
     react({